arch: migrate research strategy and finance runtimes to surface shims
